Events and others

+++ auth-service/src/main/java/com/emm/authservice/models/AuthUser.java
@@ -1,15 +1,11 @@
 package com.emm.authservice.models;
 
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 
 
 @Entity
@@ -20,7 +16,8 @@ import lombok.NoArgsConstructor;
 public class AuthUser {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String userName;
     private String password;
+
 }
